requirements.txt
flask
flask_cors
sentence-transformers
torch
transformers
google-generativeai
python-dotenv
twilio
scikit-learn
numpy
beautifulsoup4
requests
python-docx
google-generativeai 
python-docx 
pickle-mixin
pillow
pytesseract
openai
